Localization for autoformat behavior.
The new configurations autoformat.NUMERIC_REGEX and
autoformat.ALPHABETIC_REGEX are both regular expressions that validate a
*single* character, which we check before inputting it into the masked
form element.

The regular expressions in this commit are copied from
validation.instance.validateAlpha, removing spaces where they are
present. As far as I can tell, the numeric regexes were the same for all
of the languages, so those are always set to [0-9].