Work in progress on autoformat/paging interaction.
The autoformat functionality was going into an infinite loop when it was
applied to elements that weren't currently visible. This commit should get
most of the way towards fixing that. There are still some tweaks to be made
for what happens in the pageChange callback, but it no longer crashes
the browser on load.